+++ demo/src/app/components/+buttons/demo-buttons.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { ButtonsModule } from 'ngx-bootstrap/buttons';
 
@@ -17,6 +17,7 @@ import { routes } from './demo-buttons.routes';
   imports: [
     ButtonsModule.forRoot(),
     CommonModule,
+    ReactiveFormsModule,
     FormsModule,
     SharedModule,
     RouterModule.forChild(routes)