+++ scripts/typings.d.ts
@@ -8,8 +8,6 @@ declare const ENV:string;
 // google code-prettify
 declare const PR:any;
 
-declare const global:any;
-
 declare module jasmine {
   interface Matchers {
     toHaveCssClass(expected: any): boolean;