+++ src/dropdown/bs-dropdown.directive.ts
@@ -90,9 +90,15 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
   }
 
   /**
+   * Emits an event when isOpen change
+   */
+  @Output() isOpenChange: EventEmitter<any>;
+
+  /**
    * Emits an event when the popover is shown
    */
   @Output() onShown: EventEmitter<any>;
+
   /**
    * Emits an event when the popover is hidden
    */
@@ -120,10 +126,11 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
     // create dropdown component loader
     this._dropdown = this._cis
       .createLoader<BsDropdownContainerComponent>(this._elementRef, this._viewContainerRef, this._renderer)
-      .provide({provide: BsDropdownState, useValue: this._state});
+      .provide({ provide: BsDropdownState, useValue: this._state });
 
     this.onShown = this._dropdown.onShown;
     this.onHidden = this._dropdown.onHidden;
+    this.isOpenChange = this._state.isOpenChange;
 
     // set initial dropdown state from config
     this._state.autoClose = this._config.autoClose;
@@ -133,7 +140,7 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
     // fix: seems there are an issue with `routerLinkActive`
     // which result in duplicated call ngOnInit without call to ngOnDestroy
     // read more: https://github.com/valor-software/ngx-bootstrap/issues/1885
-    if (this._isInited) {return;}
+    if (this._isInited) { return; }
     this._isInited = true;
 
     this._showInline = !this.container;
@@ -157,7 +164,7 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
     // attach dropdown menu inside of dropdown
     if (this._showInline) {
       this._state.dropdownMenu
-        .then((dropdownMenu:BsComponentRef<BsDropdownMenuDirective>) => {
+        .then((dropdownMenu: BsComponentRef<BsDropdownMenuDirective>) => {
           this._inlinedMenu = dropdownMenu.viewContainer.createEmbeddedView(dropdownMenu.templateRef);
         });
     }
@@ -174,10 +181,10 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
 
     if (this._showInline) {
       this._isInlineOpen = true;
+      this.onShown.emit(true);
       this._state.isOpenChange.emit(true);
       return;
     }
-
     this._state.dropdownMenu
       .then((dropdownMenu) => {
         // check direction in which dropdown should be opened
@@ -191,7 +198,7 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
         this._dropdown
           .attach(BsDropdownContainerComponent)
           .to(this.container)
-          .position({attachment: _placement})
+          .position({ attachment: _placement })
           .show({
             content: dropdownMenu.templateRef,
             placement: _placement
@@ -199,6 +206,7 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
 
         this._state.isOpenChange.emit(true);
       });
+
   }
 
   /**
@@ -212,6 +220,7 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
 
     if (this._showInline) {
       this._isInlineOpen = false;
+      this.onHidden.emit(true);
     } else {
       this._dropdown.hide();
     }