+++ src/modal/modal.component.ts
@@ -29,6 +29,10 @@ import { ComponentLoaderFactory } from '../component-loader/component-loader.fac
 
 const TRANSITION_DURATION = 300;
 const BACKDROP_TRANSITION_DURATION = 150;
+const DISMISS_REASONS = {
+  BACKRDOP: 'backdrop-click',
+  ESC: 'esc'
+};
 
 /** Mark any code with directive to show it's content in modal */
 @Directive({
@@ -57,6 +61,8 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
 
   // seems like an Options
   public isAnimated: boolean = true;
+  /** This field contains last dismiss reason. Possible values: `backdrop-click`, `esc` and `null` (if modal was closed by direct call of `.hide()`). */
+  public dismissReason: string;
 
   public get isShown(): boolean {
     return this._isShown;
@@ -82,12 +88,14 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
   // todo: implement _dialog
   private _dialog: any;
 
+  private isNested: boolean = false;
+
   @HostListener('click', ['$event'])
   public onClick(event: any): void {
     if (this.config.ignoreBackdropClick || this.config.backdrop === 'static' || event.target !== this._element.nativeElement) {
       return;
     }
-
+    this.dismissReason = DISMISS_REASONS.BACKRDOP;
     this.hide(event);
   }
 
@@ -95,6 +103,7 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
   @HostListener('keydown.esc')
   public onEsc(): void {
     if (this.config.keyboard) {
+      this.dismissReason = DISMISS_REASONS.ESC;
       this.hide();
     }
   }
@@ -116,9 +125,11 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
 
   public ngAfterViewInit(): any {
     this._config = this._config || this.getConfig();
-    if (this._config.show) {
-      this.show();
-    }
+    setTimeout(() => {
+      if (this._config.show) {
+        this.show();
+      }
+    }, 0);
   }
 
   /* Public methods */
@@ -130,6 +141,7 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
 
   /** Allows to manually open modal */
   public show(): void {
+    this.dismissReason = null;
     this.onShow.emit(this);
     if (this._isShown) {
       return;
@@ -143,7 +155,11 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
     this.setScrollbar();
 
     if (document && document.body) {
-      this._renderer.setElementClass(document.body, ClassName.OPEN, true);
+      if (document.body.classList.contains(ClassName.OPEN)) {
+        this.isNested = true;
+      } else {
+        this._renderer.setElementClass(document.body, ClassName.OPEN, true);
+      }
     }
 
     this.showBackdrop(() => {
@@ -233,11 +249,14 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
     this._renderer.setElementAttribute(this._element.nativeElement, 'aria-hidden', 'true');
     this._renderer.setElementStyle(this._element.nativeElement, 'display', 'none');
     this.showBackdrop(() => {
-      if (document && document.body) {
-        this._renderer.setElementClass(document.body, ClassName.OPEN, false);
+      if (!this.isNested) {
+        if (document && document.body) {
+          this._renderer.setElementClass(document.body, ClassName.OPEN, false);
+        }
+        this.resetScrollbar();
       }
       this.resetAdjustments();
-      this.resetScrollbar();
+      this.focusOtherModal();
       this.onHidden.emit(this);
     });
   }
@@ -250,15 +269,9 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
       this._backdrop
         .attach(ModalBackdropComponent)
         .to('body')
-        .show({isAnimated: false});
+        .show({isAnimated: this.isAnimated});
       this.backdrop = this._backdrop._componentRef;
 
-      if (this.isAnimated) {
-        this.backdrop.instance.isAnimated = this.isAnimated;
-        Utils.reflow(this.backdrop.instance.element.nativeElement);
-      }
-
-      this.backdrop.instance.isShown = true;
       if (!callback) {
         return;
       }
@@ -319,6 +332,14 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
   // }
   // }
 
+  protected focusOtherModal() {
+    const otherOpenedModals = this._element.nativeElement.parentElement.querySelectorAll('.in[bsModal]');
+    if (!otherOpenedModals.length) {
+      return;
+    }
+    this._renderer.invokeElementMethod(otherOpenedModals[otherOpenedModals.length - 1], 'focus');
+  }
+
   /** @internal */
   protected resetAdjustments(): void {
     this._renderer.setElementStyle(this._element.nativeElement, 'paddingLeft', '');
@@ -337,17 +358,10 @@ export class ModalDirective implements AfterViewInit, OnDestroy {
       return;
     }
 
-    const fixedEl = document.querySelector(Selector.FIXED_CONTENT);
-
-    if (!fixedEl) {
-      return;
-    }
-
-    const bodyPadding = parseInt(Utils.getStyles(fixedEl).paddingRight || 0, 10);
-    this.originalBodyPadding = parseInt(document.body.style.paddingRight || 0, 10);
+    this.originalBodyPadding = parseInt(window.getComputedStyle(document.body).getPropertyValue('padding-right') || 0, 10);
 
     if (this.isBodyOverflowing) {
-      document.body.style.paddingRight = `${bodyPadding + this.scrollbarWidth}px`;
+      document.body.style.paddingRight = `${this.originalBodyPadding + this.scrollbarWidth}px`;
     }
   }